pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.68.0",
    "google-search-results>=2.4.2",
    "googlesearch-python>=1.3.0",
    "httpx>=0.24.1",
    "jinja2>=3.0.0",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.0",
    "langchain-groq>=0.0.1",
    "langgraph>=0.0.3",
    "pydantic>=2.0.0",
    "pypdf==3.16.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=0.19.0",
    "serpapi>=0.1.5",
    "sqlalchemy[asyncio]>=2.0.43",
    "sqlmodel>=0.0.8",
    "trafilatura>=1.6.1",
    "typing>=3.10.0.0",
    "uvicorn[standard]>=0.15.0",
]
